parametrize footprint avoidance
In some cases the policy to not put the obstacle that is in conflict
with robot footprint can be counter-productive. Specifically, if
the robot goes in collision with the prohibited area, instead of
failing the motion, the obstacle will be removed, allowing
the robot to enter the prohibited area.

Add the parameter that determines how to handle the conflicts
between the footprint and prohibited area.

Signed-off-by: Ilija Hadzic <ilija.hadzic@nokia-bell-labs.com>